src/requirements.txt
attrs==21.2.0
autopep8==1.6.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.2
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pyrsistent==0.18.0
python-dotenv==0.19.1
SQLAlchemy==1.4.26
toml==0.10.2
Werkzeug==2.0.2